+++ src/index.ts
@@ -1,6 +1,6 @@
 export { createBundle, getBinaryFile, getFile, getTextFile } from "./bundle.js";
 export { MarkdownBundleError } from "./error.js";
-export { formatBundleReference, resolveBundleReference } from "./paths.js";
+export { formatBundleReference, resolveBundleReference, validateBundlePath } from "./paths.js";
 export { isBinaryFile, isMarkdownFile, isTextFile } from "./types.js";
 
 export type { CreateBundleInput } from "./bundle.js";